Rename file to moonxml and drop moonscript file
For starters, moonscript is nice, but not as universal and not as fast
as pure Lua. Having to constantly synchronize the files before pushing
was also a pain, so I decided to just drop the original moonscript file
and continue working on the Lua version.

The lua file has also been renamed so it's called moonxml everywhere, no
matter if required locally or as an installed rock.