Fix & Update readme file
For whatever reason I somewhat neglected the readme file lately; this
commit updates the changelog information to properly reflect the changes
of the last few minor versions.

The version number was also removed from the "development" section, as
it's not really certain if there will be a minor version before the next
major release or not.